Split out each component in to its own file
